Added the option for Extended depth of field (CLIJ); need to optimise the radius Corrected a bug that removed first channel; if performing multichannel analysis (NA check) Macro to test different scaling factor/pixel sizes

+++ Analyse Neuron (Multi-channel) Advanced.ijm
@@ -177,9 +177,12 @@ if(cell_channel!="NA")
 	{
 		//find index of cell_channel;; keep it as string
 		idx_Hu=find_str_array(marker_no_manual,cell_channel);
-	
-		marker_names_manual=Array.deleteIndex(marker_names_manual, idx_Hu);
-		marker_no_manual=Array.deleteIndex(marker_no_manual,idx_Hu);
+		//print(idx_Hu);
+		if(idx_Hu!="NA") //if Hu found in the channel entries, delete that corresponding channel
+		{
+			marker_names_manual=Array.deleteIndex(marker_names_manual, idx_Hu);
+			marker_no_manual=Array.deleteIndex(marker_no_manual,idx_Hu);
+		}
 	}
 	cell_channel=parseInt(cell_channel);
 	if(isNaN(cell_channel)) exit("Enter channel number for cell. If leaving empty, type NA in the value");
@@ -240,20 +243,30 @@ if(sizeC>1)
 }
 
 
-//add option for extended depth of field projection for widefield images
+//added option for extended depth of field projection for widefield images
 if(sizeZ>1)
 {
 		print(img_name+" is a stack");
 		roiManager("reset");
-		waitForUser("Note the start and end of the stack.\nPress OK when done");
-		Dialog.create("Choose slices");
-  		Dialog.addNumber("Start slice", 1);
-  		Dialog.addNumber("End slice", sizeZ);
-  		Dialog.show(); 
-  		start=Dialog.getNumber();
-  		end=Dialog.getNumber();
-		run("Z Project...", "start="+start+" stop="+end+" projection=[Max Intensity]");
-		max_projection=getTitle();
+		projection_method=getBoolean("3D stack detected. Which projection method would you like?", "Maximum Intensity Projection", "Extended Depth of Field (Variance)");
+		if(projection_method==1)
+		{
+			waitForUser("Note the start and end of the stack.\nPress OK when done");
+			Dialog.create("Choose slices");
+	  		Dialog.addNumber("Start slice", 1);
+	  		Dialog.addNumber("End slice", sizeZ);
+	  		Dialog.show(); 
+	  		start=Dialog.getNumber();
+	  		end=Dialog.getNumber();
+			run("Z Project...", "start="+start+" stop="+end+" projection=[Max Intensity]");
+			max_projection=getTitle();
+			
+			
+		}
+		else 
+		{
+			max_projection=extended_depth_proj(img_name);
+		}
 }
 else 
 {
@@ -1033,4 +1046,54 @@ function draw_ganglia_outline(ganglia_img,edit_flag)
 		
 	}
 	
+}
+
+function extended_depth_proj(img)
+{
+	run("CLIJ2 Macro Extensions", "cl_device=");
+	concat_ch="";
+	selectWindow(img);
+	Stack.getDimensions(width, height, channels, slices, frames);
+	if(channels>1)
+	{
+		for(ch=1;ch<=channels;ch++)
+		{
+			selectWindow(img);
+			Stack.setChannel(ch);
+			getLut(reds, greens, blues);
+			Ext.CLIJ2_push(img);
+			radius_x = 2.0;
+			radius_y = 2.0;
+			sigma = 10.0;
+			proj_img="proj_img"+ch;
+			Ext.CLIJ2_extendedDepthOfFocusVarianceProjection(img, proj_img, radius_x, radius_y, sigma);
+			Ext.CLIJ2_pull(proj_img);
+			setLut(reds, greens, blues);
+			//Ext.CLIJ2_pull(img);
+			concat_ch=concat_ch+"c"+ch+"="+proj_img+" ";
+			
+		}
+		Ext.CLIJ2_clear();
+		//print(concat_ch);
+		run("Merge Channels...", concat_ch+" create");
+		Stack.setDisplayMode("color");
+
+	}
+	else
+	{
+		selectWindow(img);
+		getLut(reds, greens, blues);
+		Ext.CLIJ2_push(img);
+		radius_x = 2.0;
+		radius_y = 2.0;
+		sigma = 10.0;
+		proj_img="proj_img";
+		Ext.CLIJ2_extendedDepthOfFocusVarianceProjection(img, proj_img, radius_x, radius_y, sigma);
+		Ext.CLIJ2_pull(proj_img);
+		setLut(reds, greens, blues);
+	}
+	max_name="MAX_"+img;
+	rename(max_name);
+	close(img);
+	return max_name;
 }
